refactor(nkigen-lite): drop concat constant-splat path
With dead-store elimination removed, every op result now gets an HBM buffer,
so a constant read by a concat is already materialized by the time concat
runs — the splat-fill path (memset straight into the output window) is
redundant. Remove const_values and the splats plumbing; concat now loads all
inputs uniformly from hbm_map.

- direct_lower.py: drop the const_values dict and the concat closure; concat
  becomes a plain _OP_EMITTERS entry.
- direct_lower_memory.py: emit_concat loses its splats parameter and splat
  branch; _emit_splat_window deleted.

All 435 non-HW and 330 HW tensor_ir tests pass.
